Add files via upload
The word doc all the steps I took during my data analysis for a project provided to me by Coursera involving a fictional bike company and its user data. The .R doc is all the code I wrote in R for my analysis. It was meant to analyze the difference in bike rental usage between casual riders and riders who subscribed to the company and create a visualization using either R, Tableau, or some other program. I opted to do it in R as I was already using it for most of my analysis.